Move PhysicalChunkData to Chunks namespace
